backend/no-name-agent/requirements.txt
a2a-sdk
google-cloud-aiplatform
uvicorn
python-dotenv
asyncclick
gunicorn
google-adk
